refactored the textinput and its controlling state variable to a separate component
